duplicate folder removal/restructure
