+++ client/src/pages/Breakfast.js
@@ -1,0 +1,12 @@
+import React from 'react';
+function Breakfast(props){
+    return(
+        <section class="breakfast">
+            <h1>Breakfast</h1>
+            <div className='flex-row'>
+                {/* {Breakfast} */}
+            </div>
+        </section>
+    )
+}
+export default Breakfast;
