--- grandmas-cook-book/client/src/components/LoginForm.js
@@ -1,75 +1,0 @@
-import React, { useState } from 'react';
-
-import { validateEmail } from '../../utils/helpers';
-
-function Contact() {
-  const [formState, setFormState] = useState({
-    email: '',
-    password: '',
-  });
-
-  const [errorMessage, setErrorMessage] = useState('');
-  const { email, password } = formState;
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    if (!errorMessage) {
-      console.log('Logged In', formState);
-    }
-  };
-
-  const handleChange = (e) => {
-    if (e.target.email === 'email') {
-      const isValid = validateEmail(e.target.value);
-      if (!isValid) {
-        setErrorMessage('Your email is invalid.');
-      } else {
-        setErrorMessage('');
-      }
-    } else {
-      if (!e.target.value.length) {
-        setErrorMessage(`${e.target.name} is required.`);
-      } else {
-        setErrorMessage('');
-      }
-    }
-    if (!errorMessage) {
-      setFormState({ ...formState, [e.target.name]: e.target.value });
-      console.log('Handle Form', formState);
-    }
-  };
-
-  return (
-    <section>
-      <form id="contact-form" onSubmit={handleSubmit}>
-        
-        <div>
-          <label htmlFor="email">Email address:</label>
-          <input
-            type="email"
-            name="email"
-            defaultValue={email}
-            onBlur={handleChange}
-          />
-        </div>
-        <div>
-          <label htmlFor="message">Password:</label>
-          <textarea
-            name="password"
-           type='password'
-            defaultValue={password}
-            onBlur={handleChange}
-          />
-        </div>
-        {errorMessage && (
-          <div>
-            <p className="error-text">{errorMessage}</p>
-          </div>
-        )}
-        <button type="submit">Login</button>
-      </form>
-    </section>
-  );
-}
-
-export default Contact;
