--- grandmas-cook-book/client/src/pages/Dinner.js
@@ -1,12 +1,0 @@
-import React from 'react';
-function Dinner(props){
-    return(
-        <section class="dinner">
-            <h1>Dinner</h1>
-           < div className='flex-row'>
-                {/* {Dinner */}
-            </div>
-        </section>
-    )
-}
-export default Dinner;
