--- grandmas-cook-book/client/src/pages/Lunch.js
@@ -1,12 +1,0 @@
-import React from 'react';
-function Lunch(props){
-    return(
-        <section class="lunch">
-            <h1>Lunch</h1>
-            <div className='flex-row'>
-                {/* {Lunch */}
-            </div>
-        </section>
-    )
-}
-export default Lunch
