+++ server/config/connection.js
@@ -1,1 +1,8 @@
-// PLACEHOLDER FOR SERVER TO PUSH TO GITHUB
+const mongoose = require("mongoose");
+
+mongoose.connect(process.env.MONGODB_URI || "", {
+  useNewUrlParser: true,
+  useUnifiedTopology: true,
+});
+
+module.exports = mongoose.connection;
