+++ server/models/index.js
@@ -1,1 +1,3 @@
-//  placeholder
+const User = require('./User');
+
+module.exports = { User };
